rename scatterplot.py into plot.py
