+++ stackview/_plot.py
